requirements/development.txt
-r base.txt
flake8
isort
Whoosh==2.7.4
django-debug-toolbar==2.2
